catalog/tools/vertex_rag/requirements.txt
google-adk>=1.9.0
google-cloud-aiplatform>=1.50.0
